refactor(core): move seal() next to the marker it sets
`Filters.wrap()` imported `seal` from the condition helpers while the
helpers import `Filters` back, the only module cycle in the package. The
sealing logic needs nothing the collection module lacks, so it moves
there: one implementation, no cycle, no initialization-order assumption
in the bundle.
